Creation des Objet Service et route des equipes et joueur ainsi que des cotes Ajout de procédure SQL pour ajouter de nouvelles equipes, de nouveaux joueurs, et de nouvelles cotes

+++ SQL/superBowlBdd.sql
@@ -44,7 +44,7 @@ CREATE TABLE Confrontation (
 drop table if exists Cote;
 -- Création de la table Cote
 CREATE TABLE Cote (
-    coteID INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    coteID float NOT NULL AUTO_INCREMENT PRIMARY KEY,
     equipeID INT NOT NULL,
     coteVictoire DECIMAL(5,2) NOT NULL,
     FOREIGN KEY (equipeID) REFERENCES Equipe(equipeID)
